!56 【PyTorch】【dev】【NeuMF_ID0351_for_PyTorch】增加数据预处理算法+int64算子规避 * 增加数据预处理算法 * 修改数据类型为int32防止cast算子走aicpu

+++ PyTorch/dev/cv/image_classification/NeuMF_ID0351_for_PyTorch/src/data.py
@@ -123,8 +123,8 @@ class SampleGenerator(object):
                 users.append(int(row.userId))
                 items.append(int(row.negatives[i]))
                 ratings.append(float(0))  # negative samples get 0 rating
-        dataset = UserItemRatingDataset(user_tensor=torch.LongTensor(users),
-                                        item_tensor=torch.LongTensor(items),
+        dataset = UserItemRatingDataset(user_tensor=torch.IntTensor(users),
+                                        item_tensor=torch.IntTensor(items),
                                         target_tensor=torch.FloatTensor(ratings))
         return DataLoader(dataset, batch_size=batch_size, shuffle=True, drop_last=True, num_workers=16)
